requirements.txt
llama_cpp_python~=0.3.16
Flask~=3.1.3
Flask-SocketIO~=5.6.1
Werkzeug~=3.1.6
docx2txt~=0.9
faiss-cpu~=1.13.2
openpyxl~=3.1.5
PyPDF2~=3.0.1
sentence-transformers~=5.2.3